// src/app/core/model/dpl-hold-data.ts
export class DplHoldData {
  seq: number;
  addressline1: string;
  addressline2: string;
  addressline3: string;
  addressId1: string;
  addressId2: string;
  addressId3: string;
  city1: string;
  city2: string;
  city3: string;
  corpGroup: string;
  country1: string;
  country2: string;
  country3: string;
  custNameC: string;
  custNameE: string;
  custNumber: string;
  facility: string;
  ouCode: string;
  ouName: string;
  postalcode1: string;
  postalcode2: string;
  postalcode3: string;
  referenceNo: string;
  resultsStatus1: string;
  resultsStatus2: string;
  resultsStatus3: string;
  resultsUrl1: string;
  resultsUrl2: string;
  resultsUrl3: string;
  salesCode: string;
  salesName: string;
  salesDept: string;
  salesDeptName: string;
  tenant: string;
  trxDate: Date;
  trxType: string;
  docType: string;
  createdBy: string;
  formNo: string;
  owner: string;
  ownerDept: string;
  insertWhiteTemp: boolean;
  addToWhiteList1: string;
  addToWhiteList2: string;
  addToWhiteList3: string;
  addToWhiteListMsg: string;
  addToWhiteList2Msg: string;
  addToWhiteList3Msg: string;
  serviceType: string;
  custNameE1: string;
  custNameE2: string;
  custNameE3: string;
  submitDate: Date;
  status: string;
  source: string;
  remark: string;

  constructor() {
    this.seq = null;
    this.addressline1 = null;
    this.addressline2 = null;
    this.addressline3 = null;
    this.city1 = null;
    this.city2 = null;
    this.city3 = null;
    this.corpGroup = null;
    this.country1 = null;
    this.country2 = null;
    this.country3 = null;
    this.custNameC = null;
    this.custNameE = null;
    this.custNumber = null;
    this.facility = null;
    this.ouCode = null;
    this.ouName = null;
    this.postalcode1 = null;
    this.postalcode2 = null;
    this.postalcode3 = null;
    this.referenceNo = null;
    this.resultsStatus1 = null;
    this.resultsStatus2 = null;
    this.resultsStatus3 = null;
    this.resultsUrl1 = null;
    this.resultsUrl2 = null;
    this.resultsUrl3 = null;
    this.salesCode = null;
    this.salesDept = null;
    this.tenant = null;
    this.trxDate = null;
    this.trxType = null;
    this.docType = null;
    this.createdBy = null;
    this.formNo = null;
    this.owner = null;
    this.ownerDept = null;
    this.insertWhiteTemp = true;
    this.addToWhiteList1 = null;
    this.addToWhiteList2 = null;
    this.addToWhiteList3 = null;
    this.addToWhiteListMsg = null;
    this.addToWhiteList2Msg = null;
    this.addToWhiteList3Msg = null;
    this.serviceType = null;
    this.custNameE1 = null;
    this.custNameE2 = null;
    this.custNameE3 = null;
    this.submitDate = null;
    this.status = null;
    this.source = null;
    this.remark = null;
  }
}